Lightbox is now Volt
Includes some basic tests
